Make the tests fail
This commit introduces failing tests.

Rather than failing with the expected message, there is a confusing error for
the RSpec request spec.

```
NoMethodError:
  undefined method `keys' for nil:NilClass

          keys = hash.keys
```

In other repos this error can take a long time to surface, with the test
appearing to "hang", while super_diff is stuck in some sort of recursive loop
trying to build the "super" diff.